requirements.txt
fastapi
uvicorn
sqlalchemy
strawberry-graphql
pydantic
databases
aiofiles
